inventorysystem/requirements.txt
Django==3.1.7
django-autoslug==1.9.8
django-crispy-forms==1.11.2
django-nine==0.2.4
django-tinymce==3.2.0
mysqlclient==2.0.3
sqlparse==0.4.1
django-filter==2.4.0
django-cpf==0.1.0